Improvements to GameModel
